// projects/bible/mobile/src/interfaces/ChantCategoriesResponse.ts
// Generated by https://quicktype.io

export interface ChantCategoriesResponses {
  status: string;
  results: number;
  data: Data;
}

export interface Data {
  data: ICategory[];
}

export interface ICategory {
  id: number;
  uuid: string;
  name: string;
  slug: string;
  abbreviation: string;
  createdAt: string;
  updatedAt: string;
}